refactor: separate profil and profil(e) procedures into two different routers
